+++ src/load_inputs/load_resources_data.jl
@@ -34,13 +34,16 @@ function _get_policyfile_info()
     cap_res_filenames = ["Resource_capacity_reserve_margin.csv"]
     min_cap_filenames = ["Resource_minimum_capacity_requirement.csv"]
     max_cap_filenames = ["Resource_maximum_capacity_requirement.csv"]
+    h2_demand_filenames = ["Resource_hydrogen_demand.csv"]
 
     policyfile_info = (
         esr = (filenames = esr_filenames,
             setup_param = "EnergyShareRequirement"),
         cap_res = (filenames = cap_res_filenames, setup_param = "CapacityReserveMargin"),
         min_cap = (filenames = min_cap_filenames, setup_param = "MinCapReq"),
-        max_cap = (filenames = max_cap_filenames, setup_param = "MaxCapReq"))
+        max_cap = (filenames = max_cap_filenames, setup_param = "MaxCapReq"),
+        h2_demand = (
+            filenames = h2_demand_filenames, setup_param = "HydrogenMimimumProduction"))
     return policyfile_info
 end
 
@@ -123,6 +126,7 @@ function scale_vre_stor_data!(vre_stor_in::DataFrame, scale_factor::Float64)
     columns_to_scale = [:existing_cap_inverter_mw,
         :existing_cap_solar_mw,
         :existing_cap_wind_mw,
+        :existing_cap_elec_mw,
         :existing_cap_charge_dc_mw,
         :existing_cap_charge_ac_mw,
         :existing_cap_discharge_dc_mw,
@@ -133,6 +137,8 @@ function scale_vre_stor_data!(vre_stor_in::DataFrame, scale_factor::Float64)
         :max_cap_solar_mw,
         :min_cap_wind_mw,
         :max_cap_wind_mw,
+        :min_cap_elec_mw,
+        :max_cap_elec_mw,
         :min_cap_charge_ac_mw,
         :max_cap_charge_ac_mw,
         :min_cap_charge_dc_mw,
@@ -147,6 +153,8 @@ function scale_vre_stor_data!(vre_stor_in::DataFrame, scale_factor::Float64)
         :fixed_om_solar_cost_per_mwyr,
         :inv_cost_wind_per_mwyr,
         :fixed_om_wind_cost_per_mwyr,
+        :inv_cost_elec_per_mwyr,
+        :fixed_om_elec_cost_per_mwyr,
         :inv_cost_discharge_dc_per_mwyr,
         :fixed_om_cost_discharge_dc_per_mwyr,
         :inv_cost_charge_dc_per_mwyr,
@@ -164,6 +172,7 @@ function scale_vre_stor_data!(vre_stor_in::DataFrame, scale_factor::Float64)
         :min_retired_cap_inverter_mw,
         :min_retired_cap_solar_mw,
         :min_retired_cap_wind_mw,
+        :min_retired_cap_elec_mw,
         :min_retired_cap_charge_dc_mw,
         :min_retired_cap_charge_ac_mw,
         :min_retired_cap_discharge_dc_mw,
@@ -569,7 +578,7 @@ function validate_policy_dataframe!(filename::AbstractString, policy_in::DataFra
     cols = lowercase.(names(policy_in))
     filter!(col -> col ≠ "resource", cols)
 
-    accepted_cols = ["derating_factor", "esr", "esr_vrestor",
+    accepted_cols = ["derating_factor", "esr", "esr_vrestor", "h2_demand",
         [string(cap, type) for cap in ["min_cap", "max_cap"]
          for type in ("", "_stor", "_solar", "_wind")]...]
 
@@ -1137,6 +1146,9 @@ function add_resources_to_input_data!(inputs::Dict,
         # Solar PV Resources
         inputs["VS_SOLAR"] = solar(gen)
 
+        # Electrolyzer Resources
+        inputs["VS_ELEC"] = elec(gen)
+
         # DC Resources
         inputs["VS_DC"] = union(storage_dc_discharge(gen),
             storage_dc_charge(gen),
@@ -1165,6 +1177,14 @@ function add_resources_to_input_data!(inputs::Dict,
         inputs["RET_CAP_WIND"] = intersect(retirable,
             wind(gen),
             ids_with_nonneg(gen_VRE_STOR, existing_cap_wind_mw))
+        # Set of all VRE-STOR resources eligible for new electrolyzer capacity
+        inputs["NEW_CAP_ELEC"] = intersect(buildable,
+            elec(gen),
+            ids_with(gen_VRE_STOR, max_cap_elec_mw))
+        # Set of all VRE_STOR resources eligible for electrolyzer capacity retirements
+        inputs["RET_CAP_ELEC"] = intersect(retirable,
+            elec(gen),
+            ids_with_nonneg(gen_VRE_STOR, existing_cap_elec_mw))
         # Set of all VRE-STOR resources eligible for new inverter capacity
         inputs["NEW_CAP_DC"] = intersect(buildable,
             ids_with(gen_VRE_STOR, max_cap_inverter_mw),
@@ -1222,6 +1242,7 @@ function add_resources_to_input_data!(inputs::Dict,
         # Names for writing outputs
         inputs["RESOURCE_NAMES_SOLAR"] = resource_name(gen[inputs["VS_SOLAR"]])
         inputs["RESOURCE_NAMES_WIND"] = resource_name(gen[inputs["VS_WIND"]])
+        inputs["RESOURCE_NAMES_ELEC"] = resource_name(gen[inputs["VS_ELEC"]])
         inputs["RESOURCE_NAMES_DC_DISCHARGE"] = resource_name(gen[storage_dc_discharge(gen)])
         inputs["RESOURCE_NAMES_AC_DISCHARGE"] = resource_name(gen[storage_ac_discharge(gen)])
         inputs["RESOURCE_NAMES_DC_CHARGE"] = resource_name(gen[storage_dc_charge(gen)])
@@ -1229,6 +1250,7 @@ function add_resources_to_input_data!(inputs::Dict,
 
         inputs["ZONES_SOLAR"] = zone_id(gen[inputs["VS_SOLAR"]])
         inputs["ZONES_WIND"] = zone_id(gen[inputs["VS_WIND"]])
+        inputs["ZONES_ELEC"] = zone_id(gen[inputs["VS_ELEC"]])
         inputs["ZONES_DC_DISCHARGE"] = zone_id(gen[storage_dc_discharge(gen)])
         inputs["ZONES_AC_DISCHARGE"] = zone_id(gen[storage_ac_discharge(gen)])
         inputs["ZONES_DC_CHARGE"] = zone_id(gen[storage_dc_charge(gen)])